+++ lmdeploy/archs.py
@@ -137,7 +137,7 @@ def get_task(backend: str,
     """Get pipeline type and pipeline class from model config."""
     from lmdeploy.serve.core import AsyncEngine
 
-    if backend_config and getattr(backend_config, 'disable_vision_encoder', False):
+    if backend_config and backend_config.disable_vision_encoder:
         return 'llm', AsyncEngine
     _, config = get_model_arch(model_path, trust_remote_code=trust_remote_code)
     if check_vl_llm(backend, config.to_dict()):
