+++ lmdeploy/turbomind/builders/__init__.py
@@ -2,7 +2,7 @@
 """Builder sub-package — spec-driven module loading for TurboMind."""
 from __future__ import annotations
 
-from ._base import Builder, BuiltModule, SplitSide, VisualModelBuilder, _act_type_id, _torch_dtype_to_cpp
+from ._base import Builder, BuiltModule, SplitSide, _act_type_id, _torch_dtype_to_cpp
 from .attention import AttentionBuilder
 from .decoder_layer import DecoderLayerBuilder, DecoderLayerConfig
 from .deltanet import DeltaNetBuilder
@@ -13,10 +13,11 @@ from .module_list import ModuleListBuilder, ModuleListConfig
 from .moe import MoeBuilder
 from .norm import NormBuilder, make_norm_config
 from .text_model import TextModelBuilder
+from .vision_model import VisionModelBuilder
 
 __all__ = [
     # Base
-    'Builder', 'BuiltModule', 'TextModelBuilder', 'VisualModelBuilder', 'SplitSide',
+    'Builder', 'BuiltModule', 'TextModelBuilder', 'VisionModelBuilder', 'SplitSide',
     '_act_type_id', '_torch_dtype_to_cpp',
     # Builders
     'AttentionBuilder', 'FfnBuilder', 'MoeBuilder',
